requirements.txt
streamlit>=1.32.0
python-decouple>=3.8
langchain>=0.1.12
langchain-community>=0.0.27
langchain-core>=0.1.27
langchain-openai>=0.0.8
langchain-chroma>=0.0.6
chromadb>=0.4.24 